Revert "Merge branch 'master' into soapWS"
This reverts commit 5638a94c47e77dda02af6ce954abe1f5233b17a3, reversing
changes made to b2e0d6a00214bc24266fa9b2ee8e578a0d2ac1a6.

+++ persistence-modules/spring-jpa/src/main/java/org/baeldung/persistence/model/Foo.java
@@ -25,7 +25,7 @@ public class Foo implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "ID")
-    private Long id;
+    private long id;
     @Column(name = "NAME")
     private String name;
 
@@ -41,11 +41,11 @@ public class Foo implements Serializable {
         this.bar = bar;
     }
 
-    public Long getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(final Long id) {
+    public void setId(final int id) {
         this.id = id;
     }
